Call MemoryAllocator::free on dispose task
This was disabled with a comment some time ago (the comment is not
deleted to remember in case the same error is observed again). Now the
message sending has been changed for grouping so we try to enable it
again because we are running out of virtual memory.

There is another solution for this a bit less elegant; but at the moment
we haven't observed the previous issue anymore.