Use parent's task_id_t instead of TaskInfo in the TaskInfo.
This schema seems a bit complex, but it is actually more consistent to
fix the current memory leak in the TaskInfo allocation.
TaskInfo is allocated with new and many times it is never
released.

delete call is only used in taskIsBeingDeleted but sometimes the task is
not disposable (like taskfors)