Add ALLMEMORY_ACCESS_TYPE and proper upgrade table
ALLMEMORY_ACCESS_TYPE works in the same way as READWRITE_ACCESS_TYPE, except
for the upgrade rules (and potential future optimizations). This ensures that
such ordering is taken account of. It is always valid to substitute a weak
access for an allmemory access, although in some cases there will currently be
some extra serialization. Any serialization issues that arise (e.g.  due to
autowait and to enable concurrent reads on multiple nodes) will have to be
dealt with in future commits.