!eager_weak_fetch: Correctly fetch data in taskwaits if strong
A taskwait should fetch the data if it is covered by a strong task access. The
previous implementation tried to do this by adding an _isWeak field to the
bottom map entry, which was supposed to keep track of whether the task's
original access was weak. The problem was that it assumed that a bottom map
entry covers no more than one task access. In fact, no such fragmentation is
done by replaceMatchingInBottomMapLinkAndPropagate. The new implementation does
the check by iterating over the accesses inside createTaskwait (only when
cluster.eager_weak_fetch=false). This fixes allmemory_allmemory14, which was
doing the fetch when it shouldn't, and allmemory_noeagerweakfetch1, which was
not doing the fetch when it should.

Note: the bottom map entry also has an _accessType field, which has the same
problem. But it seems to be OK because _accessType is used only for limited
purposes that shouldn't cause similar problems. I added a few warnings to this
effect.  At some point we may need to fragment bottom map entries for task
accesses, but I didn't want to do this unless really necessary, because it may
affect performance. This change doesn't affect performance for the default
setting of cluster.eager_weak_fetch=true.