Taskwaits in tasks with auto fetch local data
Ensure that a taskwait in a task with a weak auto access fetches local
data (stack and lmallocs performed by the same task). The problem was that the
registerLocalAccess cannot change the type of the existing access, or whether
it was weak, since the dependency system has no logic to deal with such changes
(e.g. changing an access from weak to strong may make it enforce a dependency
when it wasn't doing so previously, but the task is already running when now it
shouldn't be). Instead we set a new flag "strong local access" in the data
access, which is checked by createTaskwait. This fixes
allmemory_allmemory{11,12,18}.