Dmallocs initially in directory
The registration of the local accesses (registerLocalAccess) for a dmalloc was
previously setting the location to be the home node. This meant that if the
first time the data is accessed is by a task on a different node (not the home
node) then there would be an unnecessary copy of the uninitialized data from
the home node. This change sets the location to the directory so the copy is
unnecessary. Nothing else should be different about the execution, as the
locality scheduler will take the initial location from the directory. This
fixes allmemory_allmemory16, which specifically tests whether this unnecessary
data fetch happens.