Stats instrumentation: fix printing of info for all nodes
Remove the call to ClusterManager::synchronizeAll(), which uses MPI_Barrier, in
the stats instrumentation shutdown function.  It is no longer valid to use any
MPI functions in the instrumentation shutdown because instrumentation shutdown
happens after the call to MPI_Finalize(), as of:

"Fix access after delete in memory shutdown."

The information is still, almost certainly, printed in the right order because of
the call to sleep().