Fix issues with WriteID
(1) Data copy step callback function calls
WriteIDManager::registerWriteIDasLocal to register the region as local. This
was lost in: "Make fetchData totally non blocking."

(2) Fix WriteIDManager's hash. The previous hash was very bad indeed, as, e.g.
there was a collision between any two regions with the same WriteID (modulo
cache size) and size if the difference in the address is a multiple of the
cache size (!)

(3) Increase the cache size from 32 to 8192 entries.