Simplify grouping and sending satisfiability messages.
Currently most of the message are sent during
processDelayedOperationsSatisfiedOriginatorsAndRemovableTasks. As we
send and receive groups of regions to release or satisfy now it is
smarter to process all of them and call
processDelayedOperationsSatisfiedOriginatorsAndRemovableTasks just at
the end to group messages as much as possible. This change affects the
information we get in the instrumentation, but we will improve that in a
latter commit.