More effective passing of WriteID
The old code created the new Write ID when a non-read-only access becomes read
satisfied. This meant that the children of a task with a weakinout would not
benefit from Write ID (except if the access was read-satisfied when the weak parent
task is scheduled and eager_weak_fetch=true). This change more carefully handles
WriteID so that it is more effective.