Allow namespace propagation until data release
The previous implementation removed the remote (offloaded to here) task's
accesses from the namespace's bottom map as soon as it was unregistered.  This
is fine with autowait and wait tasks, with no early release, since these tasks'
accesses are not unregistered until after the wait completes. But it is not
good without autowait or wait, or with early release (when implemented). This
code keeps the access on the bottom map until it is actually released (by
adding the region to the pending release accesses). It is a bit subtle in
terms of race conditions (see comments in the code).