Don't let WriteID prevent unfragmentation
The change in "Accesses cannot be merged if Write IDs are different" was
necessary to eliminate a bug, but it had the unintended side-effect of
preventing a lot of access (and taskwait) unfragmentation. This commit allows
unfragmentation of adjacent accesses whose WriteID's don't match, but in that
case it allocates a new WriteID (and registers it as local if the location is
the current node). This may in some cases reduce the effectiveness of WriteID,
but much of the time unfragmentation should be done by a wait clause (or
autowait) on the node that wrote the data, so this is hopefully OK.