Add eager send of data
Eager data send is enabled by the configuration cluster.eager_send = true.
This enables the eager sending of the data for offloaded tasks, which is done
with the satisfiability info (in the MessageTaskNew or MessageSatisfiability).
If the data is present at the offloader node, it is sent directly, otherwise a
MessageDataSend is sent to the predecessor node. The message tag for the
DataTransfer is included in the MessageDataSend and SatisfiabilityInfo. This
often helps, by reducing the latency on communication between nodes. The main
issues are (1) if a task has a weak access but part or all of it is not
accessed by any subtasks, then the data transfer would be unnecessary, and (2)
the offloading node cannot know for sure whether the recipient already has the
data.