MessageDataSend sends a vector of regions, not just one
The regions to be eagerly sent are queued and sent in single messages per
source node, at the same time as the MessageSatisfiabilitys.  In the case that
the region is already satisfied when the task is offloaded, this implementation
still sends an individual MessageDataSend per region.  But this is unlikely to
happen often in optimized programs. It is more likely that an optimized program
would offload tasks (with weak dependencies) a long time before the data
becomes ready. If it often happens that the data is already ready when the
tasks are offloaded, then it may be that the data is constant, in which case if
the data starts on another node, eager_send will probably send the same data
multiple times anyway - so eager_send wouldn't be a good optimization in that
case.