Offloaded task id is OffloadedTaskId not void *
Define a new OffloadedTaskId type for the opaque type representing the original
task on the offloader node. This cleans up a few direct casts at various places
and gives a clearer intent than just a void pointer. The value is still the
task pointer on the original node, so efficiency is the same. It should be only
a cosmetic change.