OffloadedTaskId is a globally unique counter not the pointer
It is a necessary first step to allow propagation of 'in' accesses to the same
region in multiple namespaces concurrently.  It also means that
RemoteTasksInfoMap doesn't need to use the offloader node number as part of the
key.

The cost is (1) setting the global task number, which has a low cost compared
with setting up a task and (2) a new OffloadedTasksInfoMap, which is used by
the offloader to find the original task given the OffloadedTaskId.