Fix setting of location and WriteID after taskwait
This bug has been present for some time, but it wasn't checked properly by any
of the unit tests. We now assign a new WriteID after a taskwait if the original
access is strong and not read-only. Otherwise another node may not copy any
changes made by the task following the taskwait (it may see that it already has
the old WriteID and skip the copy, therefore not getting the new changes).  We
also update the access location using the taskwait output location, if it is
set. This fixes namespacenewfail25.