Simpler solution to data release race condition
This provides a simpler solution to a race condition, by not setting the
top-level sink complete bit until it has been linked into the namespace's
bottom map. The previous solution checked the original access at step 3 and
abandoned the data release if another task gets linked in as next.

The race condition would be:
1. We make the top-level sink and set it to complete
2. We release the lock to make the previous bottom map access have the
   top-level sink as next.
3. The previous bottom map access completes, and since the top-level sink
   is complete, it triggers the data release and the top-level sink is
   discounted.
4. Since our task's original access is still on the namespace's bottom
   map, another task can be linked in as the next.
5. The next task on the namespace expects to get satisfiability in the
   namespace (and ignores "redundant" satisfiability from the parent).
   But it will never receive it so hangs.