Simplify creation of top-level sink
createTopLevelSink used to make a big effort to not create more top-level sink
fragments than necessary, by creating a single top-level sink for all
contiguous accesses that are not propagated in the namespace. It is doubtful
that it really helps because most likely the subtasks for these accesses will
complete at different times, causing the top-level sinks to become read/write
satisfied at different times, fragmenting them anyway. It also made this
function much harder to understand (and modify) than necessary, because the
top-level sink covering the previous accesses was created when the lambda
function reached an access that doesn't need one.