Add cluster.disable_remote_connect config (default true)
The cluster.disable_remote_connect option will be used to control whether to
enable direct namespace-to-namespace propagation. It is currently off by
default (actually unimplemented at the time of this commit), therefore the
value is true! This option is another "negative" one, for consistency with
cluster.disable_remote, which is confusing. We should probably change all these
at the same time.