Single MessageNoEagerSend message for whole task
It used to send one MessageNoEagerSend per finalized access, even though
MessageNoEagerSend already supported a list of regions to not eagerly send.
Now the vector of regions is part of the CPUDependencyData, and the functions
that may send a NoEagerSend message call a new function processNoEagerSendInfo
to send a single NoEagerSend message for all the accesses of the task that are
processed at the same time.