Fix rare hang in MessageDelivery at the end of the program
There was a rare hang in MessageDelivery at the end of the program, caused by
messages that had not yet been delivered when the shutdown begins. This change
makes it wait for all pending messages (and data transfers) before entering the
barrier during shutdown. This ensures that all messages have been sent before
any nodes start turning off their polling services. Also for some reason the
"Waiting for message delivery" loop requires a call to explicitly execute the
service, perhaps because the shutdown procedure is done on the LeaderThread?