Disable disableReadPropagationToNext
disableReadPropagationToNext was a misguided attempt to avoid fetching the same
data region more than once. The idea was to delay passing on the read
satisfiability in various cases when the data should have been been fetched,
but is not yet present locally. This approach more or less worked for dependent
subtasks of an offloaded task, but it is disasterous when there are successors
that will execute on another node. These successors also got delayed, which
added unnecessary serialization.

A better approach is to have a more effective implementation of WriteID (from
Use LinearRegionMap in WriteID) together with the more effective identification
of pending data accesses (LiveDataTransfers class).

The read satisfiability is passed among all successors without delay. If some
of these successors require the same data, then this will be identified, via the
two mechanisms above.