Use LinearRegionMap in WriteID
As preparation for dropping disableReadPropagationToNext, the WriteID needs to
be improved to work much better when regions get fragmented.  Instead of
storing a single region and its WriteID, we now store the regions (for a given
hashed WriteID) in a LinearRegionMap. This means that subaccesses are still
identified as present, as are larger accesses, all of whose parts are present.
We hash the WriteID before looking up in a tree in order to reduce the overhead
of locking, since different WriteIDs can be updated and accessed concurrently.