Don't create new WriteID on fragmented update
An update operation used to create a new WriteID whenever its access region
didn't exactly match the task's access's region. This is not necessary, because
the task's access region is already fragmented to the update operation.  So
either it matches exactly or it is a subaccess. In either case there is no need
to allocate a new WriteID and risk having to fetch data that is already
present.