Correct WriteID update for read satisfied (weak)concurrent
When a weakconcurrent access becomes read satisfied, it is passed a location,
which will be the correct location if the data has been updated by an earlier
weakconcurrent task in sequential order. Otherwise, the location is not
updated, since it may overwrite the correct location if already updated by the
current weakconcurrent. But even in this case it used to update the writeID (to
its old value). This was not correct, but was not found by previous tests.