For eager sends: split to message_max_size in MPIMessenger
The MPIMessenger now splits data transfers according to
cluster.message_max_size, which is needed to ensure that MPI sends and receives
are of size less than 2^31. The individual sends and receives have different
tags, otherwise they would not be ordered according to the MPI spec. This means
that the data transfers for eager sends get split correctly, with the
satisfiability info providing the first message ID in the group.
MessageId::nextMessageId now takes a parameter, which is the number of message
IDs to reserve. The individual message IDs in the group are then obtained using
MessageId::messageIdInGroup, since they are not actually consecutive values
(and if they were it would not be good practice to assume this).