Do not also fragment messages to message_max_size in workflow
Now that the MPIMessenger fragments messages to the maximum message size, it is
not necessary to also do this fragmentation in the workflow. We remove this
fragmentation. When the workflow gets the message ID for the data transfer
(MessageId::nextMessageID), it passes the number of fragments that will be
needed, which ensures that each fragment created by MPIMessenger gets a unique
tag.