Assertion that access has location on data release trigger
The access's location must be known when the data release triggers (since it is
read satisfied). There used to be a condition meaning that the data release is
ignored for any access with null (unknown) location, which didn't make sense.
This change removes the confusing condition and replaces it with an assertion.