Allow tasks to do lmallocs.
There used to be an error raised whenever any task other than main did an
lmalloc, saying that doing so was invalid without an "all_memory" access. But
this isn't true in general. It is OK to do an lmalloc inside a task if the data
is only used by the task itself and its subtasks.  It is only invalid if the
pointer is somehow returned to the parent and other tasks try to access the
data. In this case an "all_memory" access is indeed necessary.