Enable offloading of taskloops (as a whole)
The whole taskloop can now be offloaded to another node. We just need to send
the message bounds with the MessageTaskNew in exactly the same way as for a
Taskfor. The taskloop executors go through the scheduler, but they can never be
offloaded on their own. They are set as non-remote in the loop generator, just
like taskfor executors, otherwise they would inherit the remote flag from their
parent and get a data release step.