Subtask auto inherits none
Support programs like the below, which works for both cases when tasks are
inside node and also when offloaded (as dependencies are duplicated and
sent to the remote node as described in the previous commit), which is
defined to cover all memory except any regions explicitly marked as "none"
in an ancestor (e.g.  parent):

        #pragma oss task auto(...) none(A[1])
        {
                #pragma oss task auto(...)
                {
                        // A[1] is not covered by the auto, as the
                        // outer task has a none dependency.
                }
        }

A previous effort tried to remove such accesses (or parts of them) after
linking the task accesses. But this doesn't really work. For instance, the
entry that was in the bottom map before registering this task would have to be
restored, which you cannot do as an access doesn't have a link to the previous
access (only the next). It is much better to go through the accesses and remove
any parts that need removing before they are linked into the dependency system.
This means that the access is removed before there is any chance for the access
to affect anything else.