Support all constraints as allocated info in taskinfo.
This generalizes the Paul code to support all the constraints not only
the node.

* src/tasks/Task.hpp (Task) : Add new member
nanos6_task_constraints_t *_constraints.
(setConstraints) : New function to initialize the pointer.
(initConstraints) : New function to initialize the constraints.
(getConstraints) : New function to access constraints directly.
(getNode) :
(getStream) :
(getCost) : Removed functions
* src/tasks/TaskImplementation.hpp (Task) : Initialize the new values.
* src/system/ompss/AddTask.cpp (AddTask::createTask) : Allocate the
space for the constraints in the taskinfo and use the new function
setConstraints to initialize the pointer.
(AddTask::submitTask) : Use initConstraints to initialize the values.
* src/tasks/Taskloop.cpp :
* src/monitoring/TaskMonitor.cpp :
* src/scheduling/schedulers/cluster/ClusterSchedulerInterface.hpp : Use
the new API to access constraints.