Correct behaviour in various cases with missing taskwait
This fixes unit tests namespace_no_taskwait{1,2,4,5,6} and issue #9: Handling
missing taskwait at the end of main

The unregistering of a local access (stack, lmalloc or dmalloc) behaved badly
if there was a missing taskwait. Such programs are always wrong and there was
already an error meant to indicate this, but in some cases the runtime instead
gave an obscure assertion failure, hung or otherwise misbehaved.

The issue was that if there is no taskwait, then the whole region is covered by
fragments (a taskwait would have removed them). Any fragments that are not on
the bottom map (i.e. were not accessed by a subtask, but came from
fragmentation) would become removable inside unregisterLocalAccess, but
handleDataAccessStatusChanges didn't have the code to correctly remove them.
Previously it was thought that fragments could only get removed by
handleExitTaskwait. One of two things happened: (1) for a non-remote task,
there was an assertion failure, and (2) for a remote task, a loophole in the
assertion condition meant that no assertion was raised, but it tried to remove
the fragment from the accesses rather than from the fragments. The access was
still deleted, but it stayed in the access structures, leading to a
use-after-free (namespace_no_taskwait5).  Now we cleanly remove all such
fragments. This allows us to safely reach the "failIf" condition in
unregisterLocalAccess, which will complain about an incorrect program.