WriteID, MessageId and OffloadedTaskId to singleton pointers.
This is actually a preparation commit for malleability where we need a
reset function for all of them.
* src/cluster/WriteID.cpp :
* src/cluster/messages/MessageId.cpp :
* src/cluster/offloading/OffloadedTaskId.cpp : Removed sources. The new
code only initializes a singleton which is simpler to do in
src/cluster/ClusterManager.cpp.
* src/cluster/ClusterManager.cpp : Declare static singleton for the new
types.
* src/system/Bootstrap.cpp : Reorder to initialize the hardwareInfo
after ClusterManager. Taskfor executors were trying to use the
OffloadedTaskIdManager before its initialization.
*
src/cluster/offloading/OffloadedTaskId.hpp (OffloadedTaskIdManager::OffloadedTaskId)
: New type to replace OffloadedTaskId.
Fix all the occurrences.
* src/cluster/polling-services/MessageHandler.hpp : Use
OffloadedTaskIdManager::OffloadedTaskId instead of task pointer as
unordered map index. This issue was created when moved to task id
instead of task pointed and the code was not updated properly.