Implement data tracking for linear regions dependencies
Data tracking support, as needed for NUMA awareness, was previously only
implemented for the discrete dependency system. The new implementation for
linear region dependencies does not pass the NUMA node through the dependency
system (which is what the discrete dependency system does). Instead, the
function TaskDataAccesses::computeNUMAAffinity, which is called when a task
becomes ready, checks the NUMA node of each task dependency directly by calling
NUMAManager::getHomeNode. This approach is taken for simplicity, since
otherwise we would have to correctly handle merging and fragmentation of
dependencies. Since the NUMA node is not sent through the dependency system it
is not communicated among cluster nodes. This is probably the right thing to
do, because different cluster nodes may wish to put the same region on
different NUMA nodes. Another thing is that in the cluster version a task
sometimes becomes ready when the compute place is nullptr (e.g. in response
to a satisfiability message). So we cannot assume it is nullptr and use the
random engine and CPUDependencyData of the compute place (which is what the
discrete dependency system still does).