NUMA API uses lmalloc for cluster
When built with cluster support, the NUMA API functions always use
nanos6_lmalloc to allocate the memory. When the size is >= 4 KB, it happens
that the cluster memory allocator always returns 4 KB-aligned pages, so the
existing code (written for mmap) still works. We need to use madvise first,
though, because numa_interleave_memory only works for new (faulting) pages.