Dmalloc cyclic distributed across NUMA nodes
Extend Dmalloc to use a cyclic distribution across the NUMA nodes inside each
node. In the NUMAManager, we move the code that sets the NUMA affinity with the
OS from the NUMA memory allocator into a separate function,
NUMAManager::setNUMAAffinity. This allows it to be called from the NUMA data
distribution policy. It's not clear whether we should set the NUMA affinity for
data distribution at the home node only or on all nodes. At the moment it is
set only on the home nodes, and the other nodes ignore the region for the
purposes of NUMA affinity. If the config variable numa.tracking = "auto", then
data tracking is enabled by either a call to the NUMA API (as before) or a call
to nanos6_dmalloc.  Otherwise the default behaviour would be that programs
using nanos6_dmalloc but not the NUMA API would not benefit from NUMA
awareness.