Discourage work stealing from other NUMA nodes
This commit moves the tradeoff slightly towards NUMA data locality rather than
load balancing. It should probably become a configuration variable in future.
The issue is that as tasks are created (or become ready), it is possible for an
idle thread on another NUMA node to steal the task before an idle thread on the
right NUMA node has an opportunity to take it. To discourage this, an idle
thread needs to "see" the same task a given number of times (currently 10)
before it is able to steal it from another NUMA node. This small delay allows
an idle thread on the right NUMA node to take it first (which of course has no
such delay).  This needs a new function "tryReadyTask" in ReadyQueue, which
returns the next task on the ready queue. But it only removes the task from the
ready queue if it is the one that has been reserved (passed as an argument).