Remove Task::unmarkAsRemote and Task::markAsRemote
These were weird functions because tasks shouldn't change between being remote
(offloaded from another node) and not being remote.  These functions implied
that tasks spent some time in the wrong state. We instead set the right flags
when the tasks are created: remote tasks need to be created with the flag set
and taskfor collaborators and taskloop executors need to be created with the
flag clear, even if their parent (the taskfor/taskloop source) has it set.