Prepare to remove setNamespaceSuccessor
setNamespaceSuccessor is a "hack" that was needed before "Register dependencies
of offloaded tasks manually". Now that the offloaded task's dependencies are
registered directly, rather than through the Mercurium code, we are able to set
the namespace predecessor directly in the call to registerTaskDataAccess. We
add the new support and check that it would make at least as many namespace
propagations as before (there are some rare occasions when the assert fails,
e.g. in nasty/test4_executetask, but this assertion will be removed when
setNamespaceSuccessor is removed).