ClusterNodes don't have preallocated Taskfor
It doesn't make sense for ClusterNodes to have a preallocated taskfor.
Creating these preallocated taskfors meant that the runtime didn't work when
cluster.communication=disabled, since the ClusterNode was constructed before
OffloadedTaskIdManager::initialize. So when the ClusterNode created the
preallocated taskfor collaborator, the OffloadedTaskIdManager didn't exist yet,
and there was an assertion failure or segmentation fault.