Fix default grainsize for distributed taskloops
If the user does not explicitly specify the grainsize, then a default grainsize
is calculated based on the number of iterations and the number of CPUs. This
was previously done when the taskloop source was created, assuming that the
node will execute the whole iteration space. But the resulting grainsize is too
low if the taskloop is distributed across nodes. Now we defer the calculation
until the taskloop executors are created, so that the grainsize takes account
of the part of the iteration space done on the current node.