Make some ClusterMemoryManagement functions reentrant.
* src/cluster/ClusterMemoryManagement.[h,c]pp (registerDmalloc) :
construct from MessageDmalloc::DmallocMessageContent. This simplifies
code and assert that we get there all the information from the
message. We still need two different types because one serializes the
message content and the other store the pointer and makes a copy... It
should be possible to use MessageDmalloc::DmallocMessageContent as
DmallocInfo with some small changes, so we avoid having two almost exact
same types; but that comes in a next commit.
(ClusterMemoryManagement) : Convert to a class with static members.
* src/cluster/messages/MessageDmalloc.[h,c]pp (MessageDmalloc) : Remove
all wrapper functions; but add a single getter to access directly to
message content.
*
src/memory/directory/cluster/DistributionPolicy.hpp
(registerAllocation) :
(registerAllocationEqupart) : Add extra paramenter clusterSize. So use
that directly instead of calling ClusterManager::clusterSize.