Use original type when passing location from taskwait
When a taskwait fragment without a data copy (e.g. taskwait noflush or part of
a wait or autowait) completes, the location (which may have been changed by a
subtask) is passed from the taskwait to the parent task's original access. The
only exception is when the original access is a read, in which case the
original location is still valid and we don't update the location to point to
another copy elsewhere. Instead of checking whether the parent's original
access was a read, however, it used to check the access type of the taskwait
fragment. The access type of the taskwait fragment is set up from the
_accessType field of the bottom map, which is supposed to be the same as the
parent's access type (it is put there because some other parts of the code want
to check the type for heuristics like avoiding some unnecessary eager fetches).
So, despite being a bit obscure, the code was almost always working.  For all
current unit tests the two values are the same.  But there is a big warning at
the beginning of createTaskwaitFragment to say that since the bottom map
entries are not fragmented by the parent task's accesses, a single bottom map
entry could try to hold the type for multiple parent task accesses, so it is
not always correct. The problem only happens if a subtask has a single access
that covers a read and non-read access of the parent, and it updates the part
for which the parent has a non-read access.  This cannot happen with regular
access types, due to the nesting rules. But it can happen with the new
allmemory/auto access type! In any case, there is no benefit in not simply
using the correct original access type.