Add transfer base to not repeat handling code on multiple send.
This is also required in the malleability code to handle the wait. But
Was added in parts to tests the base code separately. This removes the
code duplicated in messages and data transfers

* src/cluster/messenger/TransferBase.hpp : New class to use as a base
class for all the messages.
* src/cluster/messages/Message.[h,c]pp (Message) :
* modified   src/cluster/messenger/DataTransfer.hpp (DataTransfer) :
Derive from TransferBase.

The other files just update occurrences and do small changes.