Demote allmemory of subtask to weakin if parent access is in
If a task has an allmemory/auto access but its parent has an in access, then
the allmemory access is demoted to in when the (sub)task accesses are linked
into the dependency system. This is for two reasons: (1) it is a cheap
optimization since in accesses can be done concurrently and (2) it would
otherwise be a weakinout access inside an in/weakin access, which is not
supported by the runtime (there may be an assertion failure related to
propagation of the reductionInfo). This commit fixes allmemory_true17 (which
used to fail due to the assertion failure).