Don't propagate to or from auto in the namespace
It is valid to propagate in the namespace from an in to an in and from an inout
to an inout. Propagating from at in to an inout is incorrect, as enforcing of
the WaR dependency after a group of ins is done at the offloader. It is bad to
propagate from an inout to an in because the offloader will not be able to
immediately propagate read satisfiability to any other concurrent ins that
there could be. Since an auto could turn out to be actually an in, we
cannot know whether propagating in or out of an auto access is valid. So
always disable it.