Improve MessageHandler and ClusterWorker code.
src/cluster/polling-services/ClusterWorker.hpp: Make ClusterWorker a
template class.
* src/cluster/polling-services/ClusterWorker.cpp: Deleted file static
type is a template.
* src/cluster/polling-services/MessageHandler.hpp: Reduce contention in
the lock when possible.
(_nonStealableLowPriorityMessage) : New variable for the messages that
require to be processed ad the end. Which also simplifies code.
(_waitingTaskInfoLock): New separated lock to protect access to
_waitingTaskInfo only as sometimes the normal lock does not need to be
taken and the hash table search and modification may be the most
important contention source.
(queueMessage):
(stealMessageInternal):
(notifyDoneInternal): Modify to selectively take the lock more fine
grained to reduce the contention in the workers, but also to simplify
the steal code.